pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-agent"
version = "0.1.0"
description = "Multi-agent web automation system powered by Gemini"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Your Name", email = "your.email@example.com"}]

dependencies = [
    "playwright>=1.40.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "pydantic>=2.0",
    "Pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "transformers>=4.30.0",
    "easyocr>=1.7.0",
    "opencv-python>=4.8.0",
    "google-generativeai>=0.3.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-xdist>=3.3",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "smoke: Smoke tests",
]
